requirements.txt
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python-headless==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
scipy==1.14.1
setuptools==75.6.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
stringzilla==3.11.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.12.2
uvicorn==0.34.0
